[Framework] Migrate CoinStore to coin_store module and Account CoinStores to account_coin_store module
