+++ crates/rooch-framework/doc/transaction_fee.md
@@ -16,6 +16,7 @@
 <b>use</b> <a href="">0x2::context</a>;
 <b>use</b> <a href="">0x2::object_ref</a>;
 <b>use</b> <a href="coin.md#0x3_coin">0x3::coin</a>;
+<b>use</b> <a href="coin_store.md#0x3_coin_store">0x3::coin_store</a>;
 <b>use</b> <a href="gas_coin.md#0x3_gas_coin">0x3::gas_coin</a>;
 </code></pre>
 
@@ -38,7 +39,7 @@
 
 <dl>
 <dt>
-<code>fee: <a href="_ObjectRef">object_ref::ObjectRef</a>&lt;<a href="coin.md#0x3_coin_CoinStore">coin::CoinStore</a>&gt;</code>
+<code>fee: <a href="_ObjectRef">object_ref::ObjectRef</a>&lt;<a href="coin_store.md#0x3_coin_store_CoinStore">coin_store::CoinStore</a>&gt;</code>
 </dt>
 <dd>
 
@@ -64,7 +65,7 @@
 
 
 <pre><code><b>public</b>(<b>friend</b>) <b>fun</b> <a href="transaction_fee.md#0x3_transaction_fee_genesis_init">genesis_init</a>(ctx: &<b>mut</b> Context, genesis_account: &<a href="">signer</a>)  {
-    <b>let</b> fee_store = <a href="coin.md#0x3_coin_create_coin_store">coin::create_coin_store</a>&lt;GasCoin&gt;(ctx);
+    <b>let</b> fee_store = <a href="coin_store.md#0x3_coin_store_create_coin_store">coin_store::create_coin_store</a>&lt;GasCoin&gt;(ctx);
     <a href="_global_move_to">account_storage::global_move_to</a>(ctx, genesis_account, <a href="transaction_fee.md#0x3_transaction_fee_TransactionFeePool">TransactionFeePool</a>{
         fee: fee_store,
     })
@@ -142,8 +143,8 @@ Returns the gas factor of gas.
 
 <pre><code><b>public</b>(<b>friend</b>) <b>fun</b> <a href="transaction_fee.md#0x3_transaction_fee_deposit_fee">deposit_fee</a>(ctx: &<b>mut</b> Context, <a href="gas_coin.md#0x3_gas_coin">gas_coin</a>: Coin&lt;GasCoin&gt;) {
     <b>let</b> pool = <a href="_global_borrow_mut">account_storage::global_borrow_mut</a>&lt;<a href="transaction_fee.md#0x3_transaction_fee_TransactionFeePool">TransactionFeePool</a>&gt;(ctx, @rooch_framework);
-    <b>let</b> coin_store = <a href="_borrow_mut">object_ref::borrow_mut</a>(&<b>mut</b> pool.fee);
-    <a href="coin.md#0x3_coin_deposit_to_store">coin::deposit_to_store</a>&lt;GasCoin&gt;(coin_store, <a href="gas_coin.md#0x3_gas_coin">gas_coin</a>);
+    <b>let</b> <a href="coin_store.md#0x3_coin_store">coin_store</a> = <a href="_borrow_mut">object_ref::borrow_mut</a>(&<b>mut</b> pool.fee);
+    <a href="coin_store.md#0x3_coin_store_deposit">coin_store::deposit</a>&lt;GasCoin&gt;(<a href="coin_store.md#0x3_coin_store">coin_store</a>, <a href="gas_coin.md#0x3_gas_coin">gas_coin</a>);
 }
 </code></pre>
