+++ crates/rooch-framework/sources/transaction_fee.move
@@ -6,7 +6,8 @@ module rooch_framework::transaction_fee {
     use moveos_std::account_storage;
     use moveos_std::context::Context;
     use moveos_std::object_ref::{Self, ObjectRef};
-    use rooch_framework::coin::{Self, Coin, CoinStore};
+    use rooch_framework::coin_store::{Self, CoinStore};
+    use rooch_framework::coin::Coin;
     use rooch_framework::gas_coin::{GasCoin};
 
     friend rooch_framework::genesis;
@@ -17,7 +18,7 @@ module rooch_framework::transaction_fee {
     }
 
     public(friend) fun genesis_init(ctx: &mut Context, genesis_account: &signer)  {
-        let fee_store = coin::create_coin_store<GasCoin>(ctx);
+        let fee_store = coin_store::create_coin_store<GasCoin>(ctx);
         account_storage::global_move_to(ctx, genesis_account, TransactionFeePool{
             fee: fee_store,
         })
@@ -35,7 +36,7 @@ module rooch_framework::transaction_fee {
 
     public(friend) fun deposit_fee(ctx: &mut Context, gas_coin: Coin<GasCoin>) {
         let pool = account_storage::global_borrow_mut<TransactionFeePool>(ctx, @rooch_framework);
-        let coin_store = object_ref::borrow_mut(&mut pool.fee); 
-        coin::deposit_to_store<GasCoin>(coin_store, gas_coin);
+        let coin_store = object_ref::borrow_mut(&mut pool.fee);
+        coin_store::deposit<GasCoin>(coin_store, gas_coin);
     }
 }
