+++ crates/rooch-types/src/framework/mod.rs
@@ -5,12 +5,14 @@ use crate::framework::transaction_validator::TransactionValidator;
 use moveos_types::transaction::FunctionCall;
 
 pub mod account_authentication;
+pub mod account_coin_store;
 /// Types mapping from Framework Move types to Rust types
 /// Module binding for the Framework
 //TODO there modules maybe generated by ABI in the future
 pub mod address_mapping;
 pub mod auth_validator;
 pub mod coin;
+pub mod coin_store;
 pub mod empty;
 pub mod ethereum_address;
 pub mod ethereum_light_client;
