+++ moveos/moveos-stdlib/moveos-stdlib/doc/object_ref.md
@@ -7,12 +7,12 @@
 
 -  [Resource `ObjectRef`](#0x2_object_ref_ObjectRef)
 -  [Function `new`](#0x2_object_ref_new)
--  [Function `new_with_id`](#0x2_object_ref_new_with_id)
 -  [Function `borrow`](#0x2_object_ref_borrow)
 -  [Function `borrow_mut`](#0x2_object_ref_borrow_mut)
+-  [Function `remove`](#0x2_object_ref_remove)
 -  [Function `id`](#0x2_object_ref_id)
 -  [Function `owner`](#0x2_object_ref_owner)
--  [Function `contains`](#0x2_object_ref_contains)
+-  [Function `exist_object`](#0x2_object_ref_exist_object)
 -  [Function `into_id`](#0x2_object_ref_into_id)
 
 
@@ -58,9 +58,10 @@ TODO should we support drop?
 ## Function `new`
 
 Get the object reference
+Note: new ObjectRef need the &mut Object<T>, because the ObjectRef can borrow mutable value from the object
 
 
-<pre><code><b>public</b> <b>fun</b> <a href="object_ref.md#0x2_object_ref_new">new</a>&lt;T: key&gt;(<a href="object.md#0x2_object">object</a>: &<a href="object.md#0x2_object_Object">object::Object</a>&lt;T&gt;): <a href="object_ref.md#0x2_object_ref_ObjectRef">object_ref::ObjectRef</a>&lt;T&gt;
+<pre><code><b>public</b> <b>fun</b> <a href="object_ref.md#0x2_object_ref_new">new</a>&lt;T: key&gt;(<a href="object.md#0x2_object">object</a>: &<b>mut</b> <a href="object.md#0x2_object_Object">object::Object</a>&lt;T&gt;): <a href="object_ref.md#0x2_object_ref_ObjectRef">object_ref::ObjectRef</a>&lt;T&gt;
 </code></pre>
 
 
@@ -69,36 +70,10 @@ Get the object reference
 <summary>Implementation</summary>
 
 
-<pre><code><b>public</b> <b>fun</b> <a href="object_ref.md#0x2_object_ref_new">new</a>&lt;T: key&gt;(<a href="object.md#0x2_object">object</a>: &Object&lt;T&gt;) : <a href="object_ref.md#0x2_object_ref_ObjectRef">ObjectRef</a>&lt;T&gt; {
+<pre><code><b>public</b> <b>fun</b> <a href="object_ref.md#0x2_object_ref_new">new</a>&lt;T: key&gt;(<a href="object.md#0x2_object">object</a>: &<b>mut</b> Object&lt;T&gt;) : <a href="object_ref.md#0x2_object_ref_ObjectRef">ObjectRef</a>&lt;T&gt; {
     //TODO should we track the reference count?
     <a href="object_ref.md#0x2_object_ref_ObjectRef">ObjectRef</a> {
         id: <a href="object.md#0x2_object_id">object::id</a>(<a href="object.md#0x2_object">object</a>),
-    }
-}
-</code></pre>
-
-
-
-</details>
-
-<a name="0x2_object_ref_new_with_id"></a>
-
-## Function `new_with_id`
-
-
-
-<pre><code><b>public</b>(<b>friend</b>) <b>fun</b> <a href="object_ref.md#0x2_object_ref_new_with_id">new_with_id</a>&lt;T&gt;(id: <a href="object_id.md#0x2_object_id_ObjectID">object_id::ObjectID</a>): <a href="object_ref.md#0x2_object_ref_ObjectRef">object_ref::ObjectRef</a>&lt;T&gt;
-</code></pre>
-
-
-
-<details>
-<summary>Implementation</summary>
-
-
-<pre><code><b>public</b>(<b>friend</b>) <b>fun</b> <a href="object_ref.md#0x2_object_ref_new_with_id">new_with_id</a>&lt;T&gt;(id: ObjectID): <a href="object_ref.md#0x2_object_ref_ObjectRef">ObjectRef</a>&lt;T&gt; {
-    <a href="object_ref.md#0x2_object_ref_ObjectRef">ObjectRef</a> {
-        id: id,
     }
 }
 </code></pre>
@@ -159,6 +134,34 @@ Borrow the object mutable value
 
 </details>
 
+<a name="0x2_object_ref_remove"></a>
+
+## Function `remove`
+
+Remove the object from the global storage, and return the object value
+
+
+<pre><code><b>public</b> <b>fun</b> <a href="object_ref.md#0x2_object_ref_remove">remove</a>&lt;T: key&gt;(self: <a href="object_ref.md#0x2_object_ref_ObjectRef">object_ref::ObjectRef</a>&lt;T&gt;): T
+</code></pre>
+
+
+
+<details>
+<summary>Implementation</summary>
+
+
+<pre><code><b>public</b> <b>fun</b> <a href="object_ref.md#0x2_object_ref_remove">remove</a>&lt;T: key&gt;(self: <a href="object_ref.md#0x2_object_ref_ObjectRef">ObjectRef</a>&lt;T&gt;) : T {
+    <b>let</b> <a href="object_ref.md#0x2_object_ref_ObjectRef">ObjectRef</a>{id} = self;
+    <b>let</b> <a href="object.md#0x2_object">object</a> = <a href="raw_table.md#0x2_raw_table_remove_from_global">raw_table::remove_from_global</a>(&id);
+    <b>let</b> (_id, _owner, value) = <a href="object.md#0x2_object_unpack_internal">object::unpack_internal</a>(<a href="object.md#0x2_object">object</a>);
+    value
+}
+</code></pre>
+
+
+
+</details>
+
 <a name="0x2_object_ref_id"></a>
 
 ## Function `id`
@@ -208,14 +211,14 @@ Borrow the object mutable value
 
 </details>
 
-<a name="0x2_object_ref_contains"></a>
+<a name="0x2_object_ref_exist_object"></a>
 
-## Function `contains`
+## Function `exist_object`
 
-Check if the object is still contains in the global storage
+Check if the object is still exist in the global storage
 
 
-<pre><code><b>public</b> <b>fun</b> <a href="object_ref.md#0x2_object_ref_contains">contains</a>&lt;T: key&gt;(self: &<a href="object_ref.md#0x2_object_ref_ObjectRef">object_ref::ObjectRef</a>&lt;T&gt;): bool
+<pre><code><b>public</b> <b>fun</b> <a href="object_ref.md#0x2_object_ref_exist_object">exist_object</a>&lt;T: key&gt;(self: &<a href="object_ref.md#0x2_object_ref_ObjectRef">object_ref::ObjectRef</a>&lt;T&gt;): bool
 </code></pre>
 
 
@@ -224,7 +227,7 @@ Check if the object is still contains in the global storage
 <summary>Implementation</summary>
 
 
-<pre><code><b>public</b> <b>fun</b> <a href="object_ref.md#0x2_object_ref_contains">contains</a>&lt;T: key&gt;(self: &<a href="object_ref.md#0x2_object_ref_ObjectRef">ObjectRef</a>&lt;T&gt;): bool {
+<pre><code><b>public</b> <b>fun</b> <a href="object_ref.md#0x2_object_ref_exist_object">exist_object</a>&lt;T: key&gt;(self: &<a href="object_ref.md#0x2_object_ref_ObjectRef">ObjectRef</a>&lt;T&gt;): bool {
     <a href="raw_table.md#0x2_raw_table_contains_global">raw_table::contains_global</a>(&self.id)
 }
 </code></pre>
